+++ src/scene/gsplat/unified/gsplat-manager.js
@@ -1,16 +1,13 @@
 import { Debug } from '../../../core/debug.js';
 import { Mat4 } from '../../../core/math/mat4.js';
 import { Vec3 } from '../../../core/math/vec3.js';
-import { SEMANTIC_POSITION, SEMANTIC_ATTR13, CULLFACE_NONE } from '../../../platform/graphics/constants.js';
-import { BLEND_NONE, BLEND_PREMULTIPLIED } from '../../constants.js';
 import { GraphNode } from '../../graph-node.js';
-import { ShaderMaterial } from '../../materials/shader-material.js';
-import { MeshInstance } from '../../mesh-instance.js';
 import { GSplatResourceBase } from '../gsplat-resource-base.js';
 import { GSplatCentersBuffers } from './gsplat-centers-buffer.js';
 import { GSplatInfo } from './gsplat-info.js';
 import { GSplatUnifiedSorter } from './gsplat-unified-sorter.js';
 import { GSplatWorkBuffer } from './gsplat-work-buffer.js';
+import { GSplatRenderer } from './gsplat-renderer.js';
 
 /**
  * @import { GraphicsDevice } from '../../../platform/graphics/graphics-device.js';
@@ -20,7 +17,6 @@ const cameraPosition = new Vec3();
 const cameraDirection = new Vec3();
 const translation = new Vec3();
 const invModelMat = new Mat4();
-const viewport = [0, 0];
 const tempSplats = [];
 
 /**
@@ -42,15 +38,15 @@ class GSplatManager {
     /** @type {GSplatCentersBuffers} */
     centerBuffer;
 
+    /** @type {GSplatRenderer} */
+    renderer;
+
     /**
      * An array of all splats managed by this manager.
      *
      * @type {GSplatInfo[]}
      */
     splats = [];
-
-    /** @type {MeshInstance} */
-    meshInstance;
 
     /** @type {GSplatUnifiedSorter | null} */
     sorter = null;
@@ -94,51 +90,14 @@ class GSplatManager {
         this.cameraNode = cameraNode;
         this.workBuffer = new GSplatWorkBuffer(device);
         this.centerBuffer = new GSplatCentersBuffers();
-
-        // construct the material which renders the splats from the work buffer
-        this._material = new ShaderMaterial({
-            uniqueName: 'SplatMaterial',
-            vertexGLSL: '#include "gsplatVS"',
-            fragmentGLSL: '#include "gsplatPS"',
-            vertexWGSL: '#include "gsplatVS"',
-            fragmentWGSL: '#include "gsplatPS"',
-            attributes: {
-                vertex_position: SEMANTIC_POSITION,
-                vertex_id_attrib: SEMANTIC_ATTR13
-            }
-        });
-
-        // input format
-        this._material.setDefine('GSPLAT_WORKBUFFER_DATA', true);
-
-        // input textures (work buffer textures)
-        const { workBuffer } = this;
-        this._material.setParameter('splatColor', workBuffer.colorTexture);
-        this._material.setParameter('covA', workBuffer.covATexture);
-        this._material.setParameter('covB', workBuffer.covBTexture);
-        this._material.setParameter('center', workBuffer.centerTexture);
-        this._material.setDefine('SH_BANDS', '0');
-
-        // set instance properties
-        const dither = false;
-        this._material.setParameter('numSplats', 0);
-        this._material.setParameter('splatOrder', workBuffer.orderTexture);
-        this._material.setParameter('alphaClip', 0.3);
-        this._material.setDefine(`DITHER_${dither ? 'BLUENOISE' : 'NONE'}`, '');
-        this._material.cull = CULLFACE_NONE;
-        this._material.blendType = dither ? BLEND_NONE : BLEND_PREMULTIPLIED;
-        this._material.depthWrite = !!dither;
-        this._material.update();
-
-        this.createMeshInstance();
+        this.renderer = new GSplatRenderer(device, this.node, this.workBuffer, app);
         this.createSorter();
     }
 
     destroy() {
         this.workBuffer.destroy();
         this.centerBuffer.destroy();
-        this._material.destroy();
-        this.meshInstance.destroy();
+        this.renderer.destroy();
         this.sorter.destroy();
     }
 
@@ -148,33 +107,6 @@ class GSplatManager {
         this.sorter.on('sorted', (count, version, returnCenters, orderData) => {
             this.onSorted(count, version, returnCenters, orderData);
         });
-    }
-
-    createMeshInstance() {
-
-
-        const worldLayer = this.app.scene.layers.getLayerByName('World');
-
-        if (this.meshInstance) {
-            worldLayer.removeMeshInstances([this.meshInstance]);
-            this.meshInstance.destroy();
-        }
-
-
-        const { mesh, instanceIndices } = GSplatResourceBase.createMesh(this.device, this.workBuffer.width * this.workBuffer.height);
-        this.meshInstance = new MeshInstance(mesh, this._material);
-        this.meshInstance.node = this.node;
-        this.meshInstance.setInstancing(instanceIndices, true);
-
-        // TODO: is this the best option, or do we want to generate AABB
-        // - ideally generate aabb from individual splats, update each frame
-        this.meshInstance.cull = false;
-
-        // only start rendering the splat after we've received the splat order data
-        this.meshInstance.instancingCount = 0;
-
-
-        worldLayer.addMeshInstances([this.meshInstance]);
     }
 
     add(resource, node) {
@@ -235,7 +167,7 @@ class GSplatManager {
                 this.workBuffer.resize(this.centerBuffer.textureSize);
                 this.workBuffer.setOrderData(orderData);
 
-                this.createMeshInstance();
+                this.renderer.createMeshInstance();
             }
 
             this.splats.forEach((splat) => {
@@ -249,10 +181,10 @@ class GSplatManager {
         this.workBuffer.setOrderData(orderData);
 
         // limit splat render count to exclude those behind the camera
-        this.meshInstance.instancingCount = Math.ceil(count / GSplatResourceBase.instanceSize);
+        this.renderer.meshInstance.instancingCount = Math.ceil(count / GSplatResourceBase.instanceSize);
 
         // update splat count on the material
-        this._material.setParameter('numSplats', count);
+        this.renderer._material.setParameter('numSplats', count);
     }
 
     /**
@@ -384,23 +316,6 @@ class GSplatManager {
         }
     }
 
-    updateViewport(cameraNode) {
-        const camera = cameraNode?.camera;
-        const renderTarget = camera?.renderTarget;
-        const { width, height } = renderTarget ?? this.device;
-
-        viewport[0] = width;
-        viewport[1] = height;
-
-        // adjust viewport for stereoscopic VR sessions
-        const xr = camera?.camera?.xr;
-        if (xr?.active && xr.views.list.length === 2) {
-            viewport[0] *= 0.5;
-        }
-
-        this._material.setParameter('viewport', viewport);
-    }
-
     sort(cameraNode, textureSize) {
         // Get camera's world-space properties
         const cameraMat = cameraNode.getWorldTransform();
@@ -439,7 +354,7 @@ class GSplatManager {
         });
 
         this.sorter.setSortParams(sorterRequest);
-        this.updateViewport(cameraNode);
+        this.renderer.updateViewport(cameraNode);
     }
 }
